MAP-1402 Split large test classes into smaller ones CommonData integration test and Put request tests
